BOOM Bound is a Fast-paced, Real-time, Strategic Movement and Pattern mobile game.
BOOM Bound is a Fast-paced, Real-time, Strategic Movement and Pattern
mobile game.

With thousands of seconds enjoyed (wasted?) on every single level, over
100,000 explosion in-game related deaths and 1 million in-game
explosions daily you’ll enjoy the adrenaline rush!
(NO PLAYERS WERE HARMED DURING THE PLAYING OF THIS GAME!)

The game consists of:
* 100 unique levels
* 8 unique themes
* 8 unique explosion animations

The goal of the game is to have the player maneuver his/her character
in and out of deadly explosions to get to the beacon on the other side
of the map. These levels have been refined, played and battle-tested
over and over many times giving the perfect balance of difficulty and
enjoyment.